+++ pkg/rpcclient/wsclient_test.go
@@ -381,7 +381,7 @@ func TestWSFilteredSubscriptions(t *testing.T) {
 	}{
 		{"block header primary",
 			func(t *testing.T, wsc *WSClient) {
-				primary := 3
+				primary := byte(3)
 				_, err := wsc.ReceiveHeadersOfAddedBlocks(&neorpc.BlockFilter{Primary: &primary}, make(chan *block.Header))
 				require.NoError(t, err)
 			},
@@ -389,7 +389,7 @@ func TestWSFilteredSubscriptions(t *testing.T) {
 				param := p.Value(1)
 				filt := new(neorpc.BlockFilter)
 				require.NoError(t, json.Unmarshal(param.RawMessage, filt))
-				require.Equal(t, 3, *filt.Primary)
+				require.Equal(t, byte(3), *filt.Primary)
 				require.Equal(t, (*uint32)(nil), filt.Since)
 				require.Equal(t, (*uint32)(nil), filt.Till)
 			},
@@ -404,7 +404,7 @@ func TestWSFilteredSubscriptions(t *testing.T) {
 				param := p.Value(1)
 				filt := new(neorpc.BlockFilter)
 				require.NoError(t, json.Unmarshal(param.RawMessage, filt))
-				require.Equal(t, (*int)(nil), filt.Primary)
+				require.Equal(t, (*byte)(nil), filt.Primary)
 				require.Equal(t, uint32(3), *filt.Since)
 				require.Equal(t, (*uint32)(nil), filt.Till)
 			},
@@ -419,7 +419,7 @@ func TestWSFilteredSubscriptions(t *testing.T) {
 				param := p.Value(1)
 				filt := new(neorpc.BlockFilter)
 				require.NoError(t, json.Unmarshal(param.RawMessage, filt))
-				require.Equal(t, (*int)(nil), filt.Primary)
+				require.Equal(t, (*byte)(nil), filt.Primary)
 				require.Equal(t, (*uint32)(nil), filt.Since)
 				require.Equal(t, (uint32)(3), *filt.Till)
 			},
@@ -428,7 +428,7 @@ func TestWSFilteredSubscriptions(t *testing.T) {
 			func(t *testing.T, wsc *WSClient) {
 				var (
 					since   uint32 = 3
-					primary        = 2
+					primary        = byte(2)
 					till    uint32 = 5
 				)
 				_, err := wsc.ReceiveHeadersOfAddedBlocks(&neorpc.BlockFilter{
@@ -442,14 +442,14 @@ func TestWSFilteredSubscriptions(t *testing.T) {
 				param := p.Value(1)
 				filt := new(neorpc.BlockFilter)
 				require.NoError(t, json.Unmarshal(param.RawMessage, filt))
-				require.Equal(t, 2, *filt.Primary)
+				require.Equal(t, byte(2), *filt.Primary)
 				require.Equal(t, uint32(3), *filt.Since)
 				require.Equal(t, uint32(5), *filt.Till)
 			},
 		},
 		{"blocks primary",
 			func(t *testing.T, wsc *WSClient) {
-				primary := 3
+				primary := byte(3)
 				_, err := wsc.ReceiveBlocks(&neorpc.BlockFilter{Primary: &primary}, make(chan *block.Block))
 				require.NoError(t, err)
 			},
@@ -457,7 +457,7 @@ func TestWSFilteredSubscriptions(t *testing.T) {
 				param := p.Value(1)
 				filt := new(neorpc.BlockFilter)
 				require.NoError(t, json.Unmarshal(param.RawMessage, filt))
-				require.Equal(t, 3, *filt.Primary)
+				require.Equal(t, byte(3), *filt.Primary)
 				require.Equal(t, (*uint32)(nil), filt.Since)
 				require.Equal(t, (*uint32)(nil), filt.Till)
 			},
@@ -472,7 +472,7 @@ func TestWSFilteredSubscriptions(t *testing.T) {
 				param := p.Value(1)
 				filt := new(neorpc.BlockFilter)
 				require.NoError(t, json.Unmarshal(param.RawMessage, filt))
-				require.Equal(t, (*int)(nil), filt.Primary)
+				require.Equal(t, (*byte)(nil), filt.Primary)
 				require.Equal(t, uint32(3), *filt.Since)
 				require.Equal(t, (*uint32)(nil), filt.Till)
 			},
@@ -487,7 +487,7 @@ func TestWSFilteredSubscriptions(t *testing.T) {
 				param := p.Value(1)
 				filt := new(neorpc.BlockFilter)
 				require.NoError(t, json.Unmarshal(param.RawMessage, filt))
-				require.Equal(t, (*int)(nil), filt.Primary)
+				require.Equal(t, (*byte)(nil), filt.Primary)
 				require.Equal(t, (*uint32)(nil), filt.Since)
 				require.Equal(t, (uint32)(3), *filt.Till)
 			},
@@ -496,7 +496,7 @@ func TestWSFilteredSubscriptions(t *testing.T) {
 			func(t *testing.T, wsc *WSClient) {
 				var (
 					since   uint32 = 3
-					primary        = 2
+					primary        = byte(2)
 					till    uint32 = 5
 				)
 				_, err := wsc.ReceiveBlocks(&neorpc.BlockFilter{
@@ -510,7 +510,7 @@ func TestWSFilteredSubscriptions(t *testing.T) {
 				param := p.Value(1)
 				filt := new(neorpc.BlockFilter)
 				require.NoError(t, json.Unmarshal(param.RawMessage, filt))
-				require.Equal(t, 2, *filt.Primary)
+				require.Equal(t, byte(2), *filt.Primary)
 				require.Equal(t, uint32(3), *filt.Since)
 				require.Equal(t, uint32(5), *filt.Till)
 			},
